requirements.txt
aiohttp==3.7.4.post0
alembic==1.6.5
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cmdstanpy==0.9.68
convertdate==2.3.2
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.23
dateparser==1.0.0
ephem==3.7.7.1
Flask==2.0.1
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
hijri-converter==2.1.2
holidays==0.11.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jwt==1.2.0
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.2
multidict==5.1.0
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyMeeus==0.5.11
pyparsing==2.4.7
pystan==2.19.1.1
python-binance==1.0.10
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
regex==2021.4.4
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
setuptools-git==1.2
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.17
threadpoolctl==2.1.0
tqdm==4.61.0
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.5
websockets==9.1
Werkzeug==2.0.1
xgboost==1.4.2
yarl==1.6.3
